+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `[Datagrid]` Added `disableTooltip` column setting. ([EP#8252](https://github.com/infor-design/enterprise/issues/8252))
 - `[Lookup]` Added `cssClass` setting. ([EP#8206](https://github.com/infor-design/enterprise/issues/8206))
 - `[Pager]` Added ability to set active page. ([#1571](https://github.com/infor-design/enterprise-ng/issues/1571))
+- `[Tabs]` Added `validate` setting to disable/enable error icon. ([EP#8254](https://github.com/infor-design/enterprise/issues/8254))
 
 ## 17.3.0 Fixes